refactor: extract meta Table traits and macros into meta/table.rs
Move src/meta.rs to src/meta/mod.rs and split the Table/TableEntry
traits plus the table-impl macros into a new meta/table.rs submodule.
The macros are re-exported with pub(crate) use and imported back into
mod.rs (all of them, since impl_table_traits! expands to unqualified
calls of the helper macros that must resolve at the invocation site).
No logic changes.
